Refactor out ChoiceBoxField
